requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch
torchaudio
pandas
ipython
soundfile
tqdm
PyYAML
ipykernel
matplotlib
pyaudio